+++ usvg/src/fontdb_ext.rs
@@ -1,0 +1,264 @@
+use fontdb::{ID, Database};
+use ttf_parser::GlyphId;
+
+use crate::tree;
+
+
+pub trait DatabaseExt {
+    fn populate(&mut self);
+    fn load_font(&self, id: ID) -> Option<Font>;
+    fn outline(&self, id: ID, glyph_id: GlyphId) -> Option<tree::PathData>;
+    fn has_char(&self, id: ID, c: char) -> bool;
+}
+
+impl DatabaseExt for Database {
+    fn populate(&mut self) {
+        if self.is_empty() {
+            load_system_fonts(self);
+        }
+    }
+
+    #[inline(never)]
+    fn load_font(&self, id: ID) -> Option<Font> {
+        self.with_face_data(id, |data, face_index| -> Option<Font> {
+            let font = ttf_parser::Font::from_data(data, face_index)?;
+
+            // Some fonts can have `units_per_em` set to zero, which will break out calculations.
+            // `ttf_parser` will check this for us.
+            let units_per_em = font.units_per_em()?;
+
+            let ascent = font.ascender();
+            let descent = font.descender();
+
+            let x_height = match font.x_height() {
+                Some(height) => height,
+                None => {
+                    // If not set - fallback to height * 45%.
+                    // 45% is what Firefox uses.
+                    (f32::from(ascent - descent) * 0.45) as i16
+                }
+            };
+
+            let underline = match font.underline_metrics() {
+                Some(metrics) => metrics,
+                None => {
+                    ttf_parser::LineMetrics {
+                        position: -(units_per_em as i16) / 9,
+                        thickness: units_per_em as i16 / 12,
+                    }
+                }
+            };
+
+            let line_through_position = match font.strikeout_metrics() {
+                Some(metrics) => metrics.position,
+                None => x_height / 2,
+            };
+
+            // 0.2 and 0.4 are generic offsets used by some applications (Inkscape/librsvg).
+            let mut subscript_offset = (units_per_em as f32 / 0.2).round() as i16;
+            let mut superscript_offset = (units_per_em as f32 / 0.4).round() as i16;
+            if let Some(metrics) = font.subscript_metrics() {
+                subscript_offset = metrics.y_offset;
+            }
+
+            if let Some(metrics) = font.superscript_metrics() {
+                superscript_offset = metrics.y_offset;
+            }
+
+            Some(Font {
+                id,
+                units_per_em,
+                ascent,
+                descent,
+                x_height,
+                underline_position: underline.position,
+                underline_thickness: underline.thickness,
+                line_through_position,
+                subscript_offset,
+                superscript_offset,
+            })
+        })?
+    }
+
+    #[inline(never)]
+    fn outline(&self, id: ID, glyph_id: GlyphId) -> Option<tree::PathData> {
+        self.with_face_data(id, |data, face_index| -> Option<tree::PathData> {
+            let font = ttf_parser::Font::from_data(&data, face_index)?;
+
+            let mut builder = PathBuilder { path: tree::PathData::with_capacity(16) };
+            font.outline_glyph(glyph_id, &mut builder)?;
+            Some(builder.path)
+        })?
+    }
+
+    #[inline(never)]
+    fn has_char(&self, id: ID, c: char) -> bool {
+        let res = self.with_face_data(id, |font_data, face_index| -> Option<bool> {
+            let font = ttf_parser::Font::from_data(font_data, face_index)?;
+            font.glyph_index(c)?;
+            Some(true)
+        });
+
+        match res {
+            Some(Some(true)) => true,
+            _ => false,
+        }
+    }
+}
+
+
+#[derive(Clone, Copy)]
+pub struct Font {
+    pub id: ID,
+
+    /// Guarantee to be > 0.
+    units_per_em: u16,
+
+    // All values below are in font units.
+    ascent: i16,
+    descent: i16,
+    x_height: i16,
+
+    underline_position: i16,
+    underline_thickness: i16,
+
+    // line-through thickness should be the the same as underline thickness
+    // according to the TrueType spec:
+    // https://docs.microsoft.com/en-us/typography/opentype/spec/os2#ystrikeoutsize
+    line_through_position: i16,
+
+    subscript_offset: i16,
+    superscript_offset: i16,
+}
+
+impl Font {
+    #[inline]
+    pub fn scale(&self, font_size: f64) -> f64 {
+        font_size / self.units_per_em as f64
+    }
+
+    #[inline]
+    pub fn ascent(&self, font_size: f64) -> f64 {
+        self.ascent as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn descent(&self, font_size: f64) -> f64 {
+        self.descent as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn height(&self, font_size: f64) -> f64 {
+        self.ascent(font_size) - self.descent(font_size)
+    }
+
+    #[inline]
+    pub fn x_height(&self, font_size: f64) -> f64 {
+        self.x_height as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn underline_position(&self, font_size: f64) -> f64 {
+        self.underline_position as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn underline_thickness(&self, font_size: f64) -> f64 {
+        self.underline_thickness as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn line_through_position(&self, font_size: f64) -> f64 {
+        self.line_through_position as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn subscript_offset(&self, font_size: f64) -> f64 {
+        self.subscript_offset as f64 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub fn superscript_offset(&self, font_size: f64) -> f64 {
+        self.superscript_offset as f64 * self.scale(font_size)
+    }
+}
+
+
+struct PathBuilder {
+    path: tree::PathData,
+}
+
+impl ttf_parser::OutlineBuilder for PathBuilder {
+    fn move_to(&mut self, x: f32, y: f32) {
+        self.path.push_move_to(x as f64, y as f64);
+    }
+
+    fn line_to(&mut self, x: f32, y: f32) {
+        self.path.push_line_to(x as f64, y as f64);
+    }
+
+    fn quad_to(&mut self, x1: f32, y1: f32, x: f32, y: f32) {
+        self.path.push_quad_to(
+            x1 as f64, y1 as f64,
+            x as f64, y as f64,
+        );
+    }
+
+    fn curve_to(&mut self, x1: f32, y1: f32, x2: f32, y2: f32, x: f32, y: f32) {
+        self.path.push_curve_to(
+            x1 as f64, y1 as f64,
+            x2 as f64, y2 as f64,
+            x as f64, y as f64
+        );
+    }
+
+    fn close(&mut self) {
+        self.path.push_close_path();
+    }
+}
+
+
+#[cfg(all(unix, not(target_os = "macos")))]
+fn load_system_fonts(db: &mut Database) {
+    db.load_fonts_dir("/usr/share/fonts/");
+    db.load_fonts_dir("/usr/local/share/fonts/");
+
+    if let Ok(ref home) = std::env::var("HOME") {
+        let path = std::path::Path::new(home).join(".local/share/fonts");
+        db.load_fonts_dir(path);
+    }
+
+    db.set_serif_family("Times New Roman");
+    db.set_sans_serif_family("Arial");
+    db.set_cursive_family("Comic Sans MS");
+    db.set_fantasy_family("Impact");
+    db.set_monospace_family("Courier New");
+}
+
+#[cfg(target_os = "windows")]
+fn load_system_fonts(db: &mut Database) {
+    db.load_fonts_dir("C:\\Windows\\Fonts\\");
+
+    db.set_serif_family("Times New Roman");
+    db.set_sans_serif_family("Arial");
+    db.set_cursive_family("Comic Sans MS");
+    db.set_fantasy_family("Impact");
+    db.set_monospace_family("Courier New");
+}
+
+#[cfg(target_os = "macos")]
+fn load_system_fonts(db: &mut Database) {
+    db.load_fonts_dir("/Library/Fonts");
+    db.load_fonts_dir("/System/Library/Fonts");
+
+    if let Ok(ref home) = std::env::var("HOME") {
+        let path = Path::new(home).join("Library/Fonts");
+        db.load_fonts_dir(path);
+    }
+
+    db.set_serif_family("Times New Roman");
+    db.set_sans_serif_family("Arial");
+    db.set_cursive_family("Comic Sans MS");
+    db.set_fantasy_family("Papyrus");
+    db.set_monospace_family("Courier New");
+}